Fix the project structure and compile issues: include missing sources, add the correct JUCE audio formats dependency, and resolve type/namespace problems in the core DAW files.

+++ Source/ExportManager.h
@@ -23,7 +23,7 @@ public:
         if (writer == nullptr)
             return false;
 
-        writer->writeFromAudioSampleData(buffer.getArrayOfWritePointers(), buffer.getNumChannels(), buffer.getNumSamples(), 0);
+        writer->writeFromAudioSampleData(buffer.getArrayOfWritePointers(), static_cast<int>(buffer.getNumChannels()), static_cast<int>(buffer.getNumSamples()), 0);
         return true;
     }
 };
